Renamed config/ to env/
