+++ pisek/jobs/cache.py
@@ -21,7 +21,7 @@ import yaml
 
 from pisek.utils.text import eprint
 from pisek.utils.terminal import colored_env
-from pisek.config.env import Env
+from pisek.env.env import Env
 
 CACHE_FILENAME = ".pisek_cache"
 SAVED_LAST_SIGNATURES = 5
