+++ pisek/jobs/parts/chaos_monkey.py
@@ -17,7 +17,7 @@
 import random
 import string
 
-from pisek.config.env import Env
+from pisek.env.env import Env
 from pisek.paths import TaskPath
 from pisek.jobs.parts.task_job import TaskJob
