+++ pisek/jobs/parts/checker.py
@@ -17,9 +17,9 @@
 from typing import Any, Optional
 
 from pisek.jobs.jobs import State, Job, PipelineItemFailure
-from pisek.config.env import Env
+from pisek.env.env import Env
 from pisek.paths import TaskPath
-from pisek.config.task_config import ProgramType
+from pisek.env.task_config import ProgramType
 from pisek.utils.terminal import colored_env
 from pisek.jobs.parts.task_job import TaskJobManager
 from pisek.jobs.parts.program import RunResult, RunResultKind, ProgramsJob
