+++ pisek/jobs/parts/task_job.py
@@ -23,10 +23,10 @@ import shutil
 from typing import Optional, Any, Callable, Iterable
 
 import subprocess
-from pisek.config.env import Env
-from pisek.config.task_config import ProgramLimits
+from pisek.env.env import Env
+from pisek.env.task_config import ProgramLimits
 from pisek.paths import TaskPath
-from pisek.config.task_config import SubtaskConfig, ProgramType
+from pisek.env.task_config import SubtaskConfig, ProgramType
 from pisek.jobs.jobs import Job
 from pisek.jobs.status import StatusJobManager
