+++ pisek/jobs/parts/tools.py
@@ -20,9 +20,9 @@ from typing import Optional
 
 import subprocess
 from pisek.jobs.jobs import State, Job, PipelineItemFailure
-from pisek.config.env import Env
+from pisek.env.env import Env
 from pisek.paths import TaskPath
-from pisek.config.task_config import ProgramType
+from pisek.env.task_config import ProgramType
 from pisek.jobs.parts.task_job import TaskJob, TaskJobManager
 from pisek.jobs.parts.program import ProgramsJob
