+++ pisek/self_tests/test_cms.py
@@ -5,7 +5,7 @@ import unittest
 
 from pisek.paths import GENERATED_SUBDIR
 from pisek.self_tests.util import TestFixtureVariant, overwrite_file
-from pisek.config.task_config import load_config
+from pisek.env.task_config import load_config
 
 
 class TestSumCMS(TestFixtureVariant):
