+++ pisek/self_tests/util.py
@@ -7,7 +7,7 @@ import tempfile
 import unittest
 from unittest import mock
 
-from pisek.config import config_hierarchy
+from pisek.env import config_hierarchy
 from pisek.__main__ import test_task_path
 from pisek.utils.util import quote_output, clean_task_dir
